add[Settings/JVM]: Add switch to revert to old "Execute a .jar" behaviour
Some edge cases, like jars that try to gracefully handle users using the
wrong Java version with a Swing pop-up, cause the Java version detection
to give the wrong result because the pop-up is made for Java 8. This
adds a switch to override that behaviour with the old one, from before
https://github.com/AngelAuraMC/Amethyst-Android/commit/3b1ad643427a21fea527e7252764f3c94d97c973,
which simply used the default runtime selected by the user.