fix[ModVersionListFragment]: Faulty taskProxy logic
Code thought taskProxy was being set to null, when in reality it was
actually just the value of it.

Rather than just read the value of the key, I just delete the key itself
because outside of installing something, we don't need it at all.

This bug caused anything implementing ModVersionListFragment to disable
the list after one install as the code that handles the edgecase of
taskProxy being not null was being triggered but not reset post-install.

The faulty code was intended to block the user from being able to run
multiple install tasks at once, which wouldn't go well.